requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
Mako==1.3.5
MarkupSafe==2.1.5
mysqlclient==2.2.4
numpy==1.26.4
openai==0.28.0
openpyxl==3.1.3
pandas==2.2.2
pillow==10.3.0
pydantic==2.7.2
pydantic_core==2.18.3
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
reportlab==4.2.0
requests==2.32.3
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
XlsxWriter==3.2.0
gunicorn==20.1.0
Flask-Script
<<<<<<< HEAD
Flask-Mail
=======
python-dotenv

>>>>>>> 2943027bcd9dc31cd3f716a0d516ea590f8f3f70

